data_preprocess/requirements.txt
affine==2.3.0
appdirs==1.4.4
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
cftime==1.4.1
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==1.6.0
cycler==0.10.0
dask==2021.4.0
datacube==1.8.3
decorator==4.4.2
distributed==2021.4.0
Fiona==1.8.19
fsspec==0.9.0
geocube==0.0.16
geopandas==0.9.0
greenlet==1.0.0
HeapDict==1.0.1
imageio==2.9.0
jsonschema==3.2.0
kiwisolver==1.3.1
lark-parser==0.11.2
locket==0.2.1
mahotas==1.4.11
matplotlib==3.4.1
msgpack==1.0.2
munch==2.5.0
netCDF4==1.5.6
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
openslide-python==1.1.2
pandas==1.2.3
partd==1.2.0
Pillow==8.2.0
psutil==5.8.0
psycopg2==2.8.6
pycontour==1.4.0
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
rasterio==1.2.2
rioxarray==0.3.1
Rtree==0.9.7
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
sortedcontainers==2.3.0
SQLAlchemy==1.4.7
tblib==1.7.0
tifffile==2021.4.8
toolz==0.11.1
tornado==6.1
xarray==0.17.0
zict==2.0.0
